jmux-proto: move tests into the `tests/` subfolder
Also add a few property based tests.
