+++ jmux-proto/src/lib.rs
@@ -48,7 +48,10 @@ impl fmt::Display for LocalChannelId {
     }
 }
 
-/// JMUX destination URL (<scheme>://<host>:<port>)
+/// JMUX destination URL
+///
+/// Note that this is not checking for allowed charset specified by RFC 3986 but merely validating
+/// the inner string is formatted such as: <scheme>://<host>:<port>
 #[derive(Debug, PartialEq, Eq, Clone, Hash)]
 pub struct DestinationUrl {
     inner: SmolStr,
@@ -72,21 +75,22 @@ impl DestinationUrl {
             value: s.to_owned(),
             reason: "scheme is missing",
         })?;
+        let scheme = &s[..scheme_end_idx];
+        let rest = &s[scheme_end_idx + "://".len()..];
 
-        let host_end_idx = s[scheme_end_idx..]
-            .find(':')
-            .ok_or_else(|| Error::InvalidDestinationUrl {
-                value: s.to_owned(),
-                reason: "port is missing",
-            })?
-            + scheme_end_idx;
+        let host_end_idx = rest.rfind(':').ok_or_else(|| Error::InvalidDestinationUrl {
+            value: s.to_owned(),
+            reason: "port is missing",
+        })?;
+        let host = &rest[..host_end_idx];
+        let port = &rest[host_end_idx + 1..];
 
-        let port = s[host_end_idx..].parse().map_err(|_| Error::InvalidDestinationUrl {
+        let port = port.parse().map_err(|_| Error::InvalidDestinationUrl {
             value: s.to_owned(),
             reason: "bad port",
         })?;
-        let scheme = SmolStr::new(&s[..scheme_end_idx]);
-        let host = SmolStr::new(&s[scheme_end_idx..host_end_idx]);
+        let scheme = SmolStr::new(scheme);
+        let host = SmolStr::new(host);
         let inner = SmolStr::new(s);
 
         Ok(Self {
@@ -744,214 +748,5 @@ impl ChannelClose {
         Ok(Self {
             recipient_channel_id: buf.get_u32(),
         })
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn message_type_try_from() {
-        let msg_type = MessageType::try_from(100).unwrap();
-        assert_eq!(MessageType::Open, msg_type);
-
-        let msg_type = MessageType::try_from(103).unwrap();
-        assert_eq!(MessageType::WindowAdjust, msg_type);
-
-        let msg_type = MessageType::try_from(106).unwrap();
-        assert_eq!(MessageType::Close, msg_type);
-    }
-
-    #[test]
-    fn message_type_try_err_on_invalid_bytes() {
-        let msg_type_res = MessageType::try_from(99);
-        assert!(msg_type_res.is_err());
-
-        let msg_type_res = MessageType::try_from(107);
-        assert!(msg_type_res.is_err());
-    }
-
-    #[test]
-    fn header_decode_buffer_too_short_err() {
-        let err = Header::decode(Bytes::from_static(&[])).err().unwrap();
-        assert_eq!(
-            "Not enough bytes provided to decode HEADER: received 0 bytes, expected 4 bytes",
-            err.to_string()
-        );
-    }
-
-    #[test]
-    fn header_decode() {
-        let msg = Header::decode(Bytes::from_static(&[102, 7, 16, 0])).unwrap();
-        assert_eq!(
-            Header {
-                ty: MessageType::OpenFailure,
-                size: 1808,
-                flags: 0,
-            },
-            msg
-        );
-    }
-
-    #[test]
-    fn header_encode() {
-        let header = Header {
-            ty: MessageType::OpenSuccess,
-            size: 512,
-            flags: 0,
-        };
-        let mut buf = BytesMut::new();
-        header.encode(&mut buf);
-        assert_eq!(vec![101, 2, 0, 0], buf);
-    }
-
-    fn check_encode_decode(sample_msg: Message, raw_msg: &[u8]) {
-        let mut encoded = BytesMut::new();
-        sample_msg.encode(&mut encoded).unwrap();
-        assert_eq!(raw_msg.to_vec(), encoded.to_vec());
-
-        let decoded = Message::decode(Bytes::copy_from_slice(raw_msg)).unwrap();
-        assert_eq!(sample_msg, decoded);
-    }
-
-    #[test]
-    fn channel_open() {
-        let raw_msg = &[
-            100, // msg type
-            0, 34, // msg size
-            0,  // msg flags
-            0, 0, 0, 1, // sender channel id
-            0, 0, 4, 0, // initial window size
-            4, 0, // maximum packet size
-            116, 99, 112, 58, 47, 47, 103, 111, 111, 103, 108, 101, 46, 99, 111, 109, 58, 52, 52,
-            51, // destination url: tcp://google.com:443
-        ];
-
-        let mut msg_sample = ChannelOpen::new(LocalChannelId::from(1), 4096, "tcp://google.com:443");
-        msg_sample.initial_window_size = 1024;
-        msg_sample.maximum_packet_size = 1024;
-
-        check_encode_decode(Message::Open(msg_sample), raw_msg);
-    }
-
-    #[test]
-    pub fn channel_open_success() {
-        let raw_msg = &[
-            101, // msg type
-            0, 18, // msg size
-            0,  // msg flags
-            0, 0, 0, 1, // recipient channel id
-            0, 0, 0, 2, // sender channel id
-            0, 0, 4, 0, // initial window size
-            127, 255, // maximum packet size
-        ];
-
-        let msg = ChannelOpenSuccess {
-            initial_window_size: 1024,
-            sender_channel_id: 2,
-            maximum_packet_size: 32767,
-            recipient_channel_id: 1,
-        };
-
-        check_encode_decode(Message::OpenSuccess(msg), raw_msg);
-    }
-
-    #[test]
-    pub fn channel_open_failure() {
-        let raw_msg = &[
-            102, // msg type
-            0, 17, // msg size
-            0,  // msg flags
-            0, 0, 0, 1, // recipient channel id
-            0, 0, 0, 2, // reason code
-            101, 114, 114, 111, 114, // failure description
-        ];
-
-        let msg_example = ChannelOpenFailure {
-            recipient_channel_id: 1,
-            reason_code: ReasonCode(2),
-            description: "error".to_owned(),
-        };
-
-        check_encode_decode(Message::OpenFailure(msg_example), raw_msg);
-    }
-
-    #[test]
-    pub fn channel_window_adjust() {
-        let raw_msg = &[
-            103, // msg type
-            0, 12, // msg size
-            0,  // msg flags
-            0, 0, 0, 1, // recipient channel id
-            0, 0, 2, 0, // window adjustment
-        ];
-
-        let msg_example = ChannelWindowAdjust {
-            recipient_channel_id: 1,
-            window_adjustment: 512,
-        };
-
-        check_encode_decode(Message::WindowAdjust(msg_example), raw_msg);
-    }
-
-    #[test]
-    pub fn error_on_oversized_packet() {
-        let mut buf = BytesMut::new();
-        let err = Message::data(DistantChannelId::from(1), vec![0; u16::MAX as usize])
-            .encode(&mut buf)
-            .err()
-            .unwrap();
-        assert_eq!("Packet oversized: max is 65535, got 65543", err.to_string());
-    }
-
-    #[test]
-    pub fn channel_data() {
-        let raw_msg = &[
-            104, // msg type
-            0, 12, // msg size
-            0,  // msg flags
-            0, 0, 0, 1, // recipient channel id
-            11, 12, 13, 14, // transfer data
-        ];
-
-        let msg_example = ChannelData {
-            recipient_channel_id: 1,
-            transfer_data: vec![11, 12, 13, 14],
-        };
-
-        check_encode_decode(Message::Data(msg_example), raw_msg);
-    }
-
-    #[test]
-    pub fn channel_eof() {
-        let raw_msg = &[
-            105, // msg type
-            0, 8, // msg size
-            0, // msg flags
-            0, 0, 0, 1, // recipient channel id
-        ];
-
-        let msg_example = ChannelEof {
-            recipient_channel_id: 1,
-        };
-
-        check_encode_decode(Message::Eof(msg_example), raw_msg);
-    }
-
-    #[test]
-    pub fn channel_close() {
-        let raw_msg = &[
-            106, // msg type
-            0, 8, // msg size
-            0, // msg flags
-            0, 0, 0, 1, // recipient channel id
-        ];
-
-        let msg_example = ChannelClose {
-            recipient_channel_id: 1,
-        };
-
-        check_encode_decode(Message::Close(msg_example), raw_msg);
     }
 }
