+++ jmux-proto/tests/destination_url.proptest-regressions
@@ -1,0 +1,7 @@
+# Seeds for failure cases proptest has generated in the past. It is
+# automatically read and these particular cases re-run before any
+# novel cases are generated.
+#
+# It is recommended to check this file in to source control so that
+# everyone who runs the test benefits from these saved cases.
+cc bfb965cda2c9b15c3b26208894c9e9a0a6ca90cbba9dd05dba035c69996e6522 # shrinks to scheme = " ", host = ":", port = 0
